organise tests with dirs
